requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
espn-api==0.45.1
idna==3.11
myql==1.2.7
numpy==2.3.4
oauthlib==3.3.1
pandas==2.3.3
pyaml==24.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
rauth==0.7.3
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
stringcase==1.2.0
tzdata==2025.2
urllib3==2.2.3
yahoo-oauth==2.1.1
yfpy==17.0.0
